Refactoring: Add new module, variable-response.js (VariableResponseParser)
The following went into variable-response.js:
- getVariableResponse (removed from Canned prototype)
- getEachResponse (removed from Canned prototype)
- getSelectedResponse (removed from the canned module)

canned.stripBodyComments moves to util.stripBodyComments
