+++ lib/utils.js
@@ -22,3 +22,7 @@ utils.removeJSLikeComments = function (text) {
   return text.replace(/\/\*.+?\*\/|\/\/\s.*(?=[\n\r])/g, '')
 }
 
+// replace any body comments in format //! [string]
+utils.stripBodyComments = function(data) {
+  return data && data.replace(/\/\/\! [\w]*: ([\w {}":,@./]*)/, '').trim()
+}
