+++ spec/canned.spec.js
@@ -3,14 +3,16 @@ var querystring = require("querystring")
 var canned_module = require('../canned')
 var canned = canned_module.server
 var Canned = canned_module.Canned
+var VariableResponseParser = require('../lib/variable-response')
 var path = require('path')
 var fs = require('fs')
 
-describe('Canned.prototype.getEachResponse', function(){
+
+describe('variable-response module, getEachResponse(data)', function(){
   it('splits responses on "//!"', function(done){
     var filename = './spec/test_responses/_multiple_get_responses.get.json'
     fs.readFile(filename, {encoding:'utf-8'}, function(err, data){
-      var responses = Canned.prototype.getEachResponse(data);
+      var responses = VariableResponseParser.getEachResponse(data);
       expect(responses.length).toEqual(3);
       done()
     })
@@ -19,7 +21,7 @@ describe('Canned.prototype.getEachResponse', function(){
   it('leaves the //! block on the first line', function(done){
     var filename = './spec/test_responses/_multiple_get_responses.get.json'
     fs.readFile(filename, {encoding:'utf-8'}, function(err, data){
-      var responses = Canned.prototype.getEachResponse(data);
+      var responses = VariableResponseParser.getEachResponse(data);
       for(var i=0, len=responses.length; i<len; i++){
         var lines = responses[i].split('\n')
         expect(lines[0].substr(0, 3)).toEqual('//!')
